requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
docopt==0.6.2
exceptiongroup==1.1.1
fake-useragent==1.1.3
filelock==3.10.4
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==23.3.3
fonttools==4.39.2
frozenlist==1.3.3
gast==0.4.0
gevent==22.10.2
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
markdownify==0.11.6
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
pipreqs==0.4.11
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyChatGPT==0.4.3.3
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
selenium==4.8.2
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sympy==1.11.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
torch==2.0.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
undetected-chromedriver==3.4.6
urllib3==1.26.15
webdriver-manager==4.0.0
websockets==10.4
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.8.2
zope.event==4.6
zope.interface==6.0
